CWMS-2096 - Refactored ratings controller IT into two classes, one which reuses a single store, and one which does different store options/updates
